RelationalTypeResolverInterface => ObjectType
